// libs/utils/src/lib/constants/custom-breakpoints.enum.ts
export enum CustomBreakpoints {
  Large = '(min-width: 1024px) and (max-width: 1279px)',
  Medium = '(min-width: 768px) and (max-width: 1023px)',
  Small = '(min-width: 640px) and (max-width: 767px)',
  X2Large = '(min-width: 1536px)',
  XLarge = '(min-width: 1280px) and (max-width: 1535px)',
  XSmall = '(max-width: 639px)',
}